requirements.txt
flask
flask-cors
openai
requests
mysql-connector-python
gunicorn